requirements.txt
boto3==1.8.6
botocore==1.11.6
Django==1.11.4
djangorestframework==3.8.2
docutils==0.14
jmespath==0.9.3
mysqlclient==1.3.13
python-dateutil==2.7.3
pytz==2018.5
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
